Connecting and adding function to client

+++ contracts/TodoList.sol
@@ -1,10 +1,10 @@
 pragma solidity ^0.5.0;
 
 contract TodoList {
-    uint256 public taskCount = 0;
+    uint public taskCount = 0;
 
     struct Task {
-        uint256 id;
+        uint id;
         string content;
         bool completed;
     }
